add `ui/time-slider` tests
